Make LaTeX an action, with icon `font`
Icon `calculator` is available, but since the default is text mode, I think this might be confusing.
Icon `text` is nice, but pro only
